# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wasabi Card API",
        "description": "Backend integration layer for the Wasabi Card Open API platform.\n\n## Authentication\nAll endpoints require an **X-API-KEY** header.\n\n## Response Envelope\nEvery response uses: `{ success, code, msg, data }`\n\n## Rate Limiting\n60 requests per minute per API key.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current environment"
        }
    ],
    "paths": {
...
        "/api/v1/common/regions": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Country / Region list",
                "description": "Returns all supported countries and regions (ISO 3166-1). Response is cached for 24 hours. Source: Wasabi Card /merchant/core/mcb/common/region",
                "operationId": "getRegions",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code": {
                                                        "description": "ISO 3166-1 alpha-2",
                                                        "type": "string",
                                                        "example": "SG"
                                                    },
                                                    "standardCode": {
                                                        "description": "ISO 3166-1 alpha-3",
                                                        "type": "string",
                                                        "example": "SGP"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Singapore"
                                                    },
                                                    "enableKyc": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "enableCard": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "enableGlobalTransfer": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Upstream Wasabi API error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/common/cities": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "City list",
                "description": "Returns a flat list of cities. Optionally filter by country/region code (ISO 3166-1 alpha-2). Response is cached for 24 hours. Source: Wasabi Card /merchant/core/mcb/common/city",
                "operationId": "getCities",
                "parameters": [
                    {
                        "name": "regionCode",
                        "in": "query",
                        "description": "Filter by country/region code (ISO 3166-1 alpha-2), e.g. AU",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "AU"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code": {
                                                        "description": "City code",
                                                        "type": "string",
                                                        "example": "AU_01"
                                                    },
                                                    "name": {
                                                        "description": "City name",
                                                        "type": "string",
                                                        "example": "Sydney"
                                                    },
                                                    "country": {
                                                        "description": "ISO 3166-1 alpha-2 country code",
                                                        "type": "string",
                                                        "example": "AU"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Upstream Wasabi API error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/common/mobile-codes": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Mobile Code List",
                "description": "Returns all supported international mobile dialling codes. Response is cached for 24 hours. Source: Wasabi Card /merchant/core/mcb/common/mobileAreaCode",
                "operationId": "getMobileCodes",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code": {
                                                        "description": "International dialling code (e.g. +1, +61)",
                                                        "type": "string",
                                                        "example": "+1"
                                                    },
                                                    "name": {
                                                        "description": "Country name",
                                                        "type": "string",
                                                        "example": "Canada"
                                                    },
                                                    "areaCode": {
                                                        "description": "ISO 3166-1 alpha-2 country code",
                                                        "type": "string",
                                                        "example": "CA"
                                                    },
                                                    "language": {
                                                        "description": "Default language locale",
                                                        "type": "string",
                                                        "example": "en_US"
                                                    },
                                                    "enableGlobalTransfer": {
                                                        "description": "Whether global transfer is enabled for this country",
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Upstream Wasabi API error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "code": {
                        "type": "integer",
                        "example": 401
                    },
                    "msg": {
                        "type": "string",
                        "example": "Invalid or missing API key."
                    },
                    "data": {
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API key issued to third-party developers.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}